style(lint, ios): parameter type for removeListener is double
https://github.com/facebook/react-native/blob/055ea9c7b7dea030ef16da72d1f6ecb5d95ac468/React/Modules/RCTEventEmitter.m#L119

This should quiet some compile warnings

+++ ios/RNDeviceInfo/RNDeviceInfo.m
@@ -112,7 +112,7 @@ RCT_EXPORT_METHOD(addListener : (NSString *)eventName) {
   // Keep: Required for RN built in Event Emitter Calls.
 }
 
-RCT_EXPORT_METHOD(removeListeners : (NSInteger)count) {
+RCT_EXPORT_METHOD(removeListeners : (double)count) {
   // Keep: Required for RN built in Event Emitter Calls.
 }
